% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{AddImage}
\alias{AddImage}
\alias{AddImage,Seurat,SMSeurat-method}
\title{Add image to Seurat object}
\usage{
AddImage(object, image)

\S4method{AddImage}{Seurat,SMSeurat}(object, image)
}
\arguments{
\item{object}{An object.}

\item{image}{An object that inherits from the class SpatialImage.}
}
\description{
I haven't found a function in Seurat that does this.
}
