% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran-geary.R
\name{calculateCorrelogram}
\alias{calculateCorrelogram}
\alias{calculateCorrelogram,ANY-method}
\alias{calculateCorrelogram,SpatialFeatureExperiment-method}
\alias{colGeometryCorrelogram}
\alias{colDataCorrelogram}
\alias{annotGeometryCorrelogram}
\alias{runCorrelogram}
\title{Spatial correlogram}
\usage{
\S4method{calculateCorrelogram}{ANY}(
  x,
  listw,
  order = 1,
  method = "I",
  BPPARAM = SerialParam(),
  zero.policy = NULL,
  ...
)

\S4method{calculateCorrelogram}{SpatialFeatureExperiment}(
  x,
  colGraphName,
  features,
  sample_id = NULL,
  order = 1,
  method = "I",
  exprs_values = "logcounts",
  BPPARAM = SerialParam(),
  zero.policy = NULL,
  ...
)

colGeometryCorrelogram(
  x,
  colGeometryName,
  colGraphName,
  features,
  sample_id = NULL,
  order = 1,
  method = "I",
  BPPARAM = SerialParam(),
  zero.policy = NULL,
  ...
)

colDataCorrelogram(
  x,
  colGraphName,
  features,
  sample_id = NULL,
  order = 1,
  method = "I",
  BPPARAM = SerialParam(),
  zero.policy = NULL,
  ...
)

annotGeometryCorrelogram(
  x,
  annotGeometryName,
  annotGraphName,
  features,
  sample_id = NULL,
  order = 1,
  method = "I",
  BPPARAM = SerialParam(),
  zero.policy = NULL,
  ...
)

runCorrelogram(
  sfe,
  colGraphName,
  features,
  sample_id = NULL,
  order = 1,
  method = "I",
  exprs_values = "logcounts",
  BPPARAM = SerialParam(),
  zero.policy = NULL,
  name = paste("Correlogram", method, sep = "_"),
  ...
)
}
\arguments{
\item{x}{For \code{calculateMoransI} and \code{calculateGearysC}, it can be a
numeric matrix whose rows are features/genes, or a
\code{SpatialFeatureExperiment} (SFE) object with such a matrix in an
assay.}

\item{listw}{Weighted neighborhood graph as a \code{spdep} \code{listw}
object.}

\item{order}{maximum lag order}

\item{method}{"corr" for correlation, "I" for Moran's I, "C" for Geary's C}

\item{BPPARAM}{A \code{\link{BiocParallelParam}} object specifying whether
and how computing the metric for numerous genes shall be parallelized.}

\item{zero.policy}{default NULL, use global option value; if TRUE assign zero to the lagged value of zones without neighbours, if FALSE assign NA}

\item{...}{Other arguments passed to \code{\link{sp.correlogram}}.}

\item{colGraphName}{Name of the listw graph in the SFE object that
corresponds to entities represented by columns of the gene count matrix.
Use \code{\link{colGraphNames}} to look up names of the available graphs
for cells/spots.}

\item{features}{Genes (\code{calculate*} SFE method and \code{run*}) or
numeric columns of \code{colData(x)} (\code{colData*}) or any
\code{\link{colGeometry}} (\code{colGeometryM*}) or
\code{\link{annotGeometry}} (\code{annotGeometry*}) for which the
univariate metric is to be computed. Default to \code{NULL}. When
\code{NULL}, then the metric is computed for all genes with the values in
the assay specified in the argument \code{exprs_values}. This can be
parallelized with the argument \code{BPPARAM}.}

\item{sample_id}{Sample in the SFE object whose cells/spots to use.}

\item{exprs_values}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{colGeometryName}{Name of a \code{colGeometry} \code{sf} data frame
whose numeric columns of interest are to be used to compute the metric. Use
\code{\link{colGeometryNames}} to look up names of the \code{sf} data
frames associated with cells/spots.}

\item{annotGeometryName}{Name of a \code{annotGeometry} \code{sf} data frame
whose numeric columns of interest are to be used to compute the metric. Use
\code{\link{annotGeometryNames}} to look up names of the \code{sf} data
frames associated with annotations.}

\item{annotGraphName}{Name of the listw graph in the SFE object that
corresponds to the \code{annotGeometry} of interest. Use
\code{\link{annotGraphNames}} to look up names of available annotation
graphs.}

\item{sfe}{A \code{SpatialFeatureExperiment} object.}

\item{name}{String specifying the name to be used to store the results in
\code{rowData(x)}. If not already present in the name, then the
\code{sample_id} will be appended to the name specified here separated by
an underscore.}
}
\value{
For \code{calculateCorrelogram}, a list of
  \code{spcor} objects, each element of which correslonds to a feature. For
  \code{runCorrelogram}, the \code{res} field of the \code{spcor} is taken
  and put in a list column in \code{rowData(x)}, and the SFE object with the
  new \code{rowData} is returned. For the colData, colGeometry, and
  annotGeometry versions, the results are added to an attribute of the data
  frame of interest called \code{featureData}, in a manner analogous to
  \code{rowData}.
}
\description{
Still debating whether I should write the wrapper. It should be
straightforward to call sp.correlogram directly on single colData columns.
But for genes, there's more boilerplate. I suppose, for genes, it might be
cool to compute the correlogram for a bunch of genes and plot them in the
same plot, with the error bars, or cluster them. So I'll write the wrapper.
}
