% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CreateSMSeurat}
\alias{CreateSMSeurat}
\title{Create a SMSeurat object}
\usage{
CreateSMSeurat(
  cell_centroids,
  tx_coords,
  tissue_owin = NULL,
  cell_segmentations = NULL,
  unitname = "um",
  assay = "Spatial",
  key = "spatial_"
)
}
\arguments{
\item{cell_centroids}{A data frame with 3 columns. The first column is
assumed to be the x coordinate, and the second column is assumed to be the y
coordinate. The third column is cell IDs. This is not turned into a ppp
object for ease of plotting with ggplot2, since I foresee that I'll plot cell
locations more often than transcript spots. However, this will likely change
over the development of this package after I write functions to plot various
operations on ppp objects that only have base R plotting methods in spatstat.}

\item{tx_coords}{A data frame with at least these 4 columnns: The first column
is assumed to be the x coordinate of the transcript spot, the second one is
assumed to be the y coordinate, the third one the cell this transcript
belongs to, and the fourth the gene the transcript is assigned to. All other
columns will become marks (i.e. metadata or covariate) of the points. Again,}

\item{tissue_owin}{An \code{spatstat::owin} object, or the observation window
of the tissue, such as the extent of a field of view or the boundary of the
tissue section on the slide. This is optional as this is not always provided
by published data. If omitted, and if cell_segmentations provided, then this
will be the convex hull of all cell segmentations. If cell_segmentations is
also omitted, then this will be the convex hull of all the transcript coordinates.}

\item{cell_segmentations}{A data frame with 3 columns: The first is assumed
to be the x coordinate of the segmentation boundnaries, the second is the y
coordinate, and the third the cell ID. This is optional since for published
datasets, this is not always available. When cell segmentations are unavailable,
the convex hull of the transcript points are used. See the spatstat book for
why this is problematic. The points should go counterclockwise and the first
vertex should not be repeated.}

\item{unitname}{Unit to use}

\item{assay}{Character, the assay of the Seurat object this spatial info
corresponds to.}

\item{key}{Character, must end with "_", the key Seurat use for plots.}
}
\description{
This function takes in data frames and converts the transcript locations
into ppp. The SMSeurat object is used for data from smFISH or ISS based methods
that has transcript coordinates and better yet cell segmentation polygons.
This can be used to analyze subcellular transcript localization and use the
subcellular localization to generate cell level features.
}
