% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGraphs.R
\name{findVisiumGraph}
\alias{findVisiumGraph}
\title{Find spatial neighborhood graphs for Visium spots}
\usage{
findVisiumGraph(x, sample_id, style = "W", zero.policy = NULL)
}
\arguments{
\item{x}{A \code{SpatialFeatureExperiment} object with Visium data. Column
names of the gene count matrix must be Visium barcodes, which may have a
numeric suffix to distinguish between samples (e.g. "AAACAACGAATAGTTC-1").}

\item{sample_id}{String specifying the sample, as in \code{sample_id} in
\code{colData}, for which the graph is to be constructed.}

\item{style}{\code{style} can take values \dQuote{W}, \dQuote{B}, \dQuote{C}, \dQuote{U}, \dQuote{minmax} and \dQuote{S}}

\item{zero.policy}{default NULL, use global option value; if FALSE stop with error for any empty neighbour sets, if TRUE permit the weights list to be formed with zero-length weights vectors}
}
\value{
A \code{listw} object for the neighborhood graph. The node index will
be the order of barcodes within the sample in the gene count matrix.
}
\description{
Visium spots are arranged in a hexagonal grid. This function uses the known
locations of the Visium barcodes to construct a neighborhood graph, so
adjacent spots are connected by edges. Since the known rows and columns of
the spots are used, the unit the spot centroid coordinates are in does not
matter.
}
