% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{GetSpots}
\alias{GetSpots}
\alias{GetSpots,SMSeurat-method}
\title{Get the transcript locations}
\usage{
GetSpots(object)

\S4method{GetSpots}{SMSeurat}(object)
}
\arguments{
\item{object}{An SMSeurat object.}
}
\description{
The title is self-explanatory
}
