% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{GetTissueCoordinates.SMSeurat}
\alias{GetTissueCoordinates.SMSeurat}
\title{Method for SM}
\usage{
GetTissueCoordinates.SMSeurat(object, qhulls = FALSE, rownames = TRUE)
}
\arguments{
\item{object}{The SM object.}

\item{rownames}{Logical, whether cell IDs should be rownames or a column on
its own. I hate rownames of data frames; no wonder setting rownames of tibbles
has been deprecated.}
}
\description{
Method for SM
}
