% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{mockSVGenes}
\alias{mockSVGenes}
\title{Generate SpatialExperiment instance including spatially variable genes.}
\usage{
mockSVGenes(tot_genes, de_genes, grid_size)
}
\arguments{
\item{tot_genes}{Total number of gene profiles to be generated.}

\item{de_genes}{Number of spatially variable genes.}

\item{grid_size}{Genes will be spatially arranged on a size x size grid.}
}
\value{
A \link{SpatialExperiment} object.
}
\description{
Generate SpatialExperiment instance including spatially variable genes.
}
\examples{
mockSVGenes(20, 2, 20)

}
