% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotBivariate}
\alias{plotBivariate}
\title{Plot expression of two genes or features with bivariate palette}
\usage{
plotBivariate(
  sfe,
  feature1,
  feature2,
  colGeometryName = 1L,
  sample_id = "all",
  bbox = NULL,
  palette = "BlueGold",
  size = 0.5,
  dim = 4,
  style = "fisher",
  exprs_values = "logcounts",
  swap_rownames = NULL,
  show_axes = FALSE,
  ncol = NULL
)
}
\arguments{
\item{sfe}{A \code{SpatialFeatureExperiment} object.}

\item{feature1}{First feature to plot}

\item{feature2}{Second feature to plot}

\item{colGeometryName}{Name of a \code{colGeometry} \code{sf} data frame
whose numeric columns of interest are to be used to compute the metric. Use
\code{\link{colGeometryNames}} to look up names of the \code{sf} data
frames associated with cells/spots.}

\item{sample_id}{Sample(s) in the SFE object whose cells/spots to use. Can be
"all" to compute metric for all samples; the metric is computed separately
for each sample.}

\item{bbox}{A bounding box to specify a smaller region to plot, useful when
the dataset is large. Can be a named numeric vector with names "xmin",
"xmax", "ymin", and "ymax", in any order. If plotting multiple samples, it
should be a matrix with sample IDs as column names and "xmin", "ymin",
"xmax", and "ymax" as row names. If multiple samples are plotted but
\code{bbox} is a vector rather than a matrix, then the same bounding box
will be used for all samples. You may see points at the edge of the
geometries if the intersection between the bounding box and a geometry
happens to be a point there. If \code{NULL}, then the entire tissue is
plotted.}

\item{palette}{Name of bivariate palette; see \code{\link[biscale]{bi_pal}}
for complete list of built-in palettes in the `biscale` package.}

\item{size}{Fixed size of points. For points defaults to 0.5. Ignored if
\code{size_by} is specified.}

\item{dim}{The dimensions of the palette. To use the built-in palettes,
    this value must be either \code{2}, \code{3}, or \code{4}. A value of
    \code{3}, for example, would be used to create a three-by-three bivariate
    map with a total of 9 classes.

    If you are using a custom palette, this value may be larger (though these
    maps can be very hard to interpret).

    If you are using pre-made factors, both factors must have the same number
    of levels as this value.}

\item{style}{A string identifying the style used to calculate breaks. Currently
    supported styles are \code{"quantile"}, \code{"equal"}, \code{"fisher"},
    and \code{"jenks"}. If both \code{x} and \code{y} are factors, this argument can
    be omitted.

    Note that older versions of \code{biscale} used \code{"quantile"} as the default
    for this argument. Now that \code{bi_class} accepts factors, this argument
    no longer as a default and older code will error.}

\item{exprs_values}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{swap_rownames}{Column name of \code{rowData(object)} to be used to
identify features instead of \code{rownames(object)} when labeling plot
elements. If not found in \code{rowData}, then rownames of the gene count
matrix will be used.}

\item{show_axes}{Logical, whether to show axes.}

\item{ncol}{Number of columns if plotting multiple features. Defaults to
\code{NULL}, which means using the same logic as \code{facet_wrap}, which
is used by \code{patchwork}'s \code{\link{wrap_plots}} by default.}
}
\description{
This function uses the `biscale` package to plot two features simultaneously
using a bivariate palette. The features can come from the assays or colData.
}
\note{
Legend must be plotted separated as another `ggplot` object. See
\code{\link[biscale]{bi_legend}}.
}
\concept{Spatial plotting}
