% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotColGraph}
\alias{plotColGraph}
\alias{plotAnnotGraph}
\title{Plot spatial graphs}
\usage{
plotColGraph(
  sfe,
  colGraphName,
  colGeometryName = NULL,
  sample_id = NULL,
  segment_size = 0.5,
  geometry_size = 0.5
)

plotAnnotGraph(
  sfe,
  annotGraphName,
  annotGeometryName,
  sample_id = NULL,
  segment_size = 0.5,
  geometry_size = 0.5
)
}
\arguments{
\item{sfe}{A \code{SpatialFeatureExperiment} object.}

\item{colGraphName}{Name of graph associated with columns of the gene count
matrix.}

\item{colGeometryName}{Name of a \code{colGeometry} \code{sf} data frame
whose numeric columns of interest are to be used to compute the metric. Use
\code{\link{colGeometryNames}} to look up names of the \code{sf} data
frames associated with cells/spots.}

\item{sample_id}{Sample in the SFE object whose cells/spots to use.}

\item{annotGeometryName}{Name of a \code{annotGeometry} of the SFE object, to
annotate the gene expression plot.}
}
\value{
A ggplot2 object.
}
\description{
A ggplot version of \code{spdep::plot.nb}, reducing boilerplate for SFE
objects.
}
