% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_spec.R
\name{set_last_spec}
\alias{set_last_spec}
\title{Set the spec to be fetched by last_spec()}
\usage{
set_last_spec(value)
}
\description{
Set the spec to be fetched by last_spec()
}
\seealso{
[last_spec()]
}
\keyword{internal}
