% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{SMSeurat-class}
\alias{SMSeurat-class}
\alias{SMSeurat}
\title{S4 class for SpatialImage with transcript locations and cell segmentations}
\description{
The SMSeurat class inherits from the \code{SpatialImage} class from Seurat, and
extends it to smFISH based or ISS based datasets where transcript locations
and/or cell segmentation polygons are available.
}
\section{Slots}{

\describe{
\item{\code{coordinates}}{A data frame with 3 columns: "x", and "y" for
the x and y coordinates, "cell" for cell IDs.}

\item{\code{tissue_owin}}{A list of one \code{spatstat::owin} object, or the
observation window of the tissue.}

\item{\code{qhulls}}{A list whose length is the same as the number of rows of
coordinates, for cell segmentation polygons.}

\item{\code{spots}}{A list whose length is the same as the number of rows of
coordinates, for transcript coordinates. Each element is a ppp.}

\item{\code{unitname}}{Name of the unit of the coordinates.}
}}

