% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{SpatialPlot2}
\alias{SpatialPlot2}
\title{My alternative to Seurat's plotting function}
\usage{
SpatialPlot2(object, group.by = NULL)
}
\arguments{
\item{object}{A Seurat object}

\item{group.by}{Same as in Seurat's SpatialPlot.}
}
\description{
Because at present, Seurat's version is not extensible, though at present,
my version is simpler. For now it only colors by the metadata. It would also
only work on SMSeurat.
}
