% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VoyageR.R
\name{VoyageR}
\alias{VoyageR}
\title{Launch Voyager as a Shiny Gadget}
\usage{
VoyageR(data = NULL, browser = FALSE, envir = .GlobalEnv)
}
\arguments{
\item{data}{Dataset to load}

\item{browser}{Open in an external browser}

\item{envir}{Environment to use (defaults to the global environment)}
}
\value{
If a visualisation has been specified its JSON will be returned as a string
}
\description{
Launch Voyager as a Shiny Gadget
}
\examples{
library(dplyr)

data("mtcars")
mtcars \%>\%
 VoyageR() \%>\%
 vegawidget::as_vegaspec()
}
