% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran-geary.R
\name{calculateMoranPlot}
\alias{calculateMoranPlot}
\alias{calculateMoranPlot,ANY-method}
\alias{calculateMoranPlot,SpatialFeatureExperiment-method}
\alias{colDataMoranPlot}
\alias{colGeometryMoranPlot}
\alias{annotGeometryMoranPlot}
\alias{runMoranPlot}
\title{Moran scatterplot}
\usage{
\S4method{calculateMoranPlot}{ANY}(x, listw, BPPARAM = SerialParam(), zero.policy = NULL, ...)

\S4method{calculateMoranPlot}{SpatialFeatureExperiment}(
  x,
  colGraphName,
  features = NULL,
  sample_id = NULL,
  exprs_values = "logcounts",
  BPPARAM = SerialParam(),
  zero.policy = NULL,
  ...
)

colDataMoranPlot(
  x,
  colGraphName,
  features,
  sample_id = NULL,
  BPPARAM = SerialParam(),
  zero.policy = NULL,
  ...
)

colGeometryMoranPlot(
  x,
  colGeometryName,
  colGraphName,
  features,
  sample_id = NULL,
  BPPARAM = SerialParam(),
  zero.policy = NULL,
  ...
)

annotGeometryMoranPlot(
  x,
  annotGeometryName,
  annotGraphName,
  features,
  sample_id = NULL,
  BPPARAM = SerialParam(),
  zero.policy = NULL,
  ...
)

runMoranPlot(
  x,
  colGraphName,
  features,
  sample_id = NULL,
  exprs_values = "logcounts",
  BPPARAM = SerialParam(),
  zero.policy = NULL,
  name = "MoranPlot",
  ...
)
}
\arguments{
\item{x}{For \code{calculateMoransI} and \code{calculateGearysC}, it can be a
numeric matrix whose rows are features/genes, or a
\code{SpatialFeatureExperiment} (SFE) object with such a matrix in an
assay. For \code{runMoransI} and \code{runGearysC}, and the \code{colData},
\code{colGeometry}, and \code{annotGeometry} versions, it must be a
\code{SpatialFeatureExperiment} object.}

\item{listw}{Weighted neighborhood graph as a \code{spdep} \code{listw}
object.}

\item{BPPARAM}{A \code{\link{BiocParallelParam}} object specifying whether
and how computing the metric for numerous genes shall be parallelized.}

\item{zero.policy}{default NULL, use global option value; if TRUE assign zero to the lagged value of zones without neighbours, if FALSE assign NA}

\item{...}{Other arguments passed to \code{\link{moran.plot}}.}

\item{colGraphName}{Name of the listw graph in the SFE object that
corresponds to entities represented by columns of the gene count matrix.
Use \code{\link{colGraphNames}} to look up names of the available graphs
for cells/spots.}

\item{features}{Genes (\code{calculate*} SFE method and \code{run*}) or
numeric columns of \code{colData(x)} (\code{colData*}) or any
\code{\link{colGeometry}} (\code{colGeometryM*}) or
\code{\link{annotGeometry}} (\code{annotGeometry*}) for which the
univariate metric is to be computed. Default to \code{NULL}. When
\code{NULL}, then the metric is computed for all genes with the values in
the assay specified in the argument \code{exprs_values}. This can be
parallelized with the argument \code{BPPARAM}.}

\item{sample_id}{Sample in the SFE object whose cells/spots to use.}

\item{exprs_values}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{colGeometryName}{Name of a \code{colGeometry} \code{sf} data frame
whose numeric columns of interest are to be used to compute the metric. Use
\code{\link{colGeometryNames}} to look up names of the \code{sf} data
frames associated with cells/spots.}

\item{annotGeometryName}{Name of a \code{annotGeometry} \code{sf} data frame
whose numeric columns of interest are to be used to compute the metric. Use
\code{\link{annotGeometryNames}} to look up names of the \code{sf} data
frames associated with annotations.}

\item{annotGraphName}{Name of the listw graph in the SFE object that
corresponds to the \code{annotGeometry} of interest. Use
\code{\link{annotGraphNames}} to look up names of available annotation
graphs.}

\item{name}{String specifying the name to be used to store the results in
\code{rowData(x)}. If not already present in the name, then the
\code{sample_id} will be appended to the name specified here separated by
an underscore.}
}
\value{
For \code{calculateMoranPlot}, a list of data frames that are the
  output of \code{\link{moran.plot}}. The names of the list are the names of
  the features. For \code{runMoranPlot}, the list is added to a column in
  \code{rowData(x)}. For the colData, colGeometry, and annotGeometry
  versions, the results are added to an attribute of the data frame of
  interest called \code{featureData}, in a manner analogous to
  \code{rowData}. The plot is not made.
}
\description{
To make it easier to compute the Moran scatterplot for a bunch of genes at
once. And to reduce the boilerplate for getting the gene expression values.
Burning question: how shall I compare these across genes? Also to organize
the results, which was one of the motivations of SFE.
}
