% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran-geary.R
\name{calculateMoransI}
\alias{calculateMoransI}
\alias{calculateGearysC}
\alias{calculateMoransI,ANY-method}
\alias{calculateGearysC,ANY-method}
\alias{calculateMoransI,SpatialFeatureExperiment-method}
\alias{calculateGearysC,SpatialFeatureExperiment-method}
\alias{colDataMoransI}
\alias{colDataGearysC}
\alias{colGeometryMoransI}
\alias{colGeometryGearysC}
\alias{annotGeometryMoransI}
\alias{annotGeometryGearysC}
\alias{runMoransI}
\alias{runGearysC}
\title{Calculate univariate spatial autocorrelation}
\usage{
\S4method{calculateMoransI}{ANY}(x, listw, BPPARAM = SerialParam(), zero.policy = NULL)

\S4method{calculateGearysC}{ANY}(x, listw, BPPARAM = SerialParam(), zero.policy = NULL)

\S4method{calculateMoransI}{SpatialFeatureExperiment}(
  x,
  colGraphName,
  features,
  sample_id = NULL,
  exprs_values = "logcounts",
  BPPARAM = SerialParam(),
  zero.policy = NULL,
  ...
)

\S4method{calculateGearysC}{SpatialFeatureExperiment}(
  x,
  colGraphName,
  features,
  sample_id = NULL,
  exprs_values = "logcounts",
  BPPARAM = SerialParam(),
  zero.policy = NULL,
  ...
)

colDataMoransI(
  x,
  colGraphName,
  features,
  sample_id = NULL,
  BPPARAM = SerialParam(),
  zero.policy = NULL,
  ...
)

colDataGearysC(
  x,
  colGraphName,
  features,
  sample_id = NULL,
  BPPARAM = SerialParam(),
  zero.policy = NULL,
  ...
)

colGeometryMoransI(
  x,
  colGeometryName,
  colGraphName,
  features,
  sample_id = NULL,
  BPPARAM = SerialParam(),
  zero.policy = NULL,
  ...
)

colGeometryGearysC(
  x,
  colGeometryName,
  colGraphName,
  features,
  sample_id = NULL,
  BPPARAM = SerialParam(),
  zero.policy = NULL,
  ...
)

annotGeometryMoransI(
  x,
  annotGeometryName,
  annotGraphName,
  features,
  sample_id = NULL,
  BPPARAM = SerialParam(),
  zero.policy = NULL,
  ...
)

annotGeometryGearysC(
  x,
  annotGeometryName,
  annotGraphName,
  features,
  sample_id = NULL,
  BPPARAM = SerialParam(),
  zero.policy = NULL,
  ...
)

runMoransI(
  x,
  colGraphName,
  features,
  sample_id = NULL,
  exprs_values = "logcounts",
  BPPARAM = SerialParam(),
  zero.policy = NULL,
  name = "MoransI"
)

runGearysC(
  x,
  colGraphName,
  features,
  sample_id = NULL,
  exprs_values = "logcounts",
  BPPARAM = SerialParam(),
  zero.policy = NULL,
  name = "MoransI"
)
}
\arguments{
\item{x}{For \code{calculateMoransI} and \code{calculateGearysC}, it can be a
numeric matrix whose rows are features/genes, or a
\code{SpatialFeatureExperiment} (SFE) object with such a matrix in an
assay. For \code{runMoransI} and \code{runGearysC}, and the \code{colData},
\code{colGeometry}, and \code{annotGeometry} versions, it must be a
\code{SpatialFeatureExperiment} object.}

\item{listw}{Weighted neighborhood graph as a \code{spdep} \code{listw}
object.}

\item{BPPARAM}{A \code{\link{BiocParallelParam}} object specifying whether
and how computing Moran's I or Geary's C for numerous genes shall be
parallelized.}

\item{zero.policy}{default NULL, use global option value; if TRUE assign zero to the lagged value of zones without neighbours, if FALSE assign NA}

\item{colGraphName}{Name of the listw graph in the SFE object that
corresponds to entities represented by columns of the gene count matrix.
Use \code{\link{colGraphNames}} to look up names of the available graphs
for cells/spots.}

\item{features}{Genes (\code{calculateMoransI/GearysC} SFE method and
\code{runMoransI/GearysC}) or numeric columns of \code{colData(x)}
(\code{colDataMoransI/GearysC}) or any \code{\link{colGeometry}}
(\code{colGeometryMoransI/GearysC}) or \code{\link{annotGeometry}}
(\code{annotGeometryMoransI/GearysC}) for which Moran's I or Geary's C is
to be computed. Default to \code{NULL}. When \code{NULL}, then Moran's I or
Geary's C is computed for all genes with the values in the assay specified
in the argument \code{exprs_values}. This can be parallelized with the
argument \code{BPPARAM}.}

\item{sample_id}{Sample in the SFE object whose cells/spots to use to
calculate Moran's I or Geary's C.}

\item{exprs_values}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{colGeometryName}{Name of a \code{colGeometry} \code{sf} data frame
whose numeric columns of interest are to be used to compute Moran's I or
Geary's C. Use \code{\link{colGeometryNames}} to look up names of the
\code{sf} data frames associated with cells/spots.}

\item{annotGeometryName}{Name of a \code{annotGeometry} \code{sf} data frame
whose numeric columns of interest are to be used to compute Moran's I or
Geary's C. Use \code{\link{annotGeometryNames}} to look up names of the
\code{sf} data frames associated with annotations.}

\item{annotGraphName}{Name of the listw graph in the SFE object that
corresponds to the \code{annotGeometry} of interest. Use
\code{\link{annotGraphNames}} to look up names of available annotation
graphs.}

\item{name}{String specifying the name to be used to store the results in
\code{rowData(x)} for \code{runMoransI} and \code{runGearysC}. If the SFE
object has more than one \code{sample_id}, then the \code{sample_id} will
be appended to the name specified here separated by an underscore.}
}
\value{
For \code{calculate*} and the \code{colData}, \code{colGeometry}, and
  \code{annotGeometry} versions, a \code{DataFrame} with two columns: The
  first one is I for Moran's I or C for Geary's C, and the second one is K
  for sample kurtosis. For \code{run*}, a \code{SpatialFeatureExperiment}
  object with the Moran's I or Geary's C values added to a column of
  \code{rowData(x)}, whose name is specified in the \code{name} argument,
  with \code{sample_id} appended if applicable.
}
\description{
Compute Moran's I or Geary's C on gene expression or numeric columns of
colData, colGeometry, or annotGeometry of a \code{SpatialFeatureExperiment}
object. Multithreading is supported when computing for numerous genes.
}
