% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{calculateMultivariate}
\alias{calculateMultivariate}
\alias{calculateMultivariate,ANY,SFEMethod-method}
\alias{calculateMultivariate,ANY,character-method}
\alias{calculateMultivariate,SpatialFeatureExperiment,ANY-method}
\alias{runMultivariate}
\title{Multivariate spatial data analysis}
\usage{
\S4method{calculateMultivariate}{ANY,SFEMethod}(x, type, listw = NULL, ...)

\S4method{calculateMultivariate}{ANY,character}(x, type, listw = NULL, ...)

\S4method{calculateMultivariate}{SpatialFeatureExperiment,ANY}(
  x,
  type,
  colGraphName = 1L,
  ntop = 500,
  subset_row = NULL,
  exprs_values = "logcounts",
  sample_action = c("joint", "separate"),
  BPPARAM = SerialParam(),
  ...
)

runMultivariate(
  x,
  type,
  colGraphName = 1L,
  ntop = 500,
  subset_row = NULL,
  exprs_values = "logcounts",
  sample_action = c("joint", "separate"),
  BPPARAM = SerialParam(),
  name = NULL,
  dest = c("reducedDim", "colData"),
  ...
)
}
\arguments{
\item{x}{A numeric matrix whose rows are features/genes, or a
\code{SpatialFeatureExperiment} (SFE) object with such a matrix in an
assay.}

\item{type}{A string, must be one of the following: moran, geary, moran.test,
geary.test, moran.mc, geary.mc, sp.mantel.mc, globalG.test, sp.correlogram,
localmoran, localmoran_perm, localC, localC_perm, localG, localG_perm,
LOSH, LOSH.mc, LOSH.cs, and moran.plot. See \code{spdep} documentation for
the corresponding functions for method specific arguments. Can also be an
\code{\link{SFEMethod}} object, or a string matching the name of an
\code{SFEMethod} object. The methods mentioned above correspond to
\code{SFEMethod} objects already implemented in the Voyager package. You
can implement new \code{SFEMethod} objects to apply Voyager functions to
other spatial analysis methods. This is in part inspired by the
\code{caret}, \code{parsnip}, and \code{BiocSingular} packages.}

\item{listw}{Weighted neighborhood graph as a \code{spdep} \code{listw}
object.}

\item{...}{Extra arguments passed to the specific multivariate method. For
example, see \code{\link{multispati_rsp}} for arguments for MULTISPATI PCA.}

\item{colGraphName}{Name of the listw graph in the SFE object that
corresponds to entities represented by columns of the gene count matrix.
Use \code{\link{colGraphNames}} to look up names of the available graphs
for cells/spots. Note that for multiple \code{sample_id}s, it is assumed
that all of them have a graph of this same name.}

\item{ntop}{Numeric scalar specifying the number of features with the highest variances to use for dimensionality reduction.}

\item{subset_row}{Vector specifying the subset of features to use for dimensionality reduction.
This can be a character vector of row names, an integer vector of row indices or a logical vector.}

\item{exprs_values}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{sample_action}{Character, either "joint" or "separate". Spatial methods
depend on the spatial coordinates and/or spatial neighborhood graph, which
is why \code{SpatialExperiment} uses \code{sample_id} to keep coordinates
from different samples separate. Some spatial methods can be sensibly run
jointly for multiple samples. In this case, "joint" will run the method
jointly for all samples, and "separate" will run the method separately for
each sample and concatenate the results. If the output of each sample is a
3D array, as in GWPCA, then the \code{abind} package needs to be installed
to concatenate the results of the different samples.}

\item{BPPARAM}{A \code{\link{BiocParallelParam}} object specifying whether
and how computing the metric for numerous genes shall be parallelized. This
is to parallelize computation across multiple samples when there are a
large number of samples. Be cautious if using an optimized BLAS for matrix
operations that supports multithreading.}

\item{name}{Name to use to store the results, defaults to the name in the
\code{SFEMethod} object passed to argument \code{type}. Can be set to
distinguish between results from the same method but with different
parameters.}

\item{dest}{Character, either "reducedDim" or "colData". If the output of the
multivariate method is a matrix or array, as in spatially informed
dimension reduction, then the only option is "reducedDim", so the results
will be stored in \code{\link{reducedDim}} of the SFE object. If the output
is a vector, as in the multivariate version of \code{\link{localC}}, then
choosing "colData" will store the output in a new column in
\code{colData(x)}.}
}
\value{
In \code{calculateMultivariate}, a matrix for cell embeddings whose
  attributes include loadings and eigenvalues if relevant, ready to be added
  to the SFE object with \code{reducedDim} setter. For \code{run*}, a
  \code{SpatialFeatureExperiment} object with the results added. See Details
  for where the results are stored.
}
\description{
These functions perform multivariate spatial data analysis, usually spatially
informed dimension reduction.
}
\examples{
# example code
library(SFEData)
library(scater)
library(scran)
sfe <- McKellarMuscleData()
sfe <- logNormCounts(sfe)
gvs <- modelGeneVar(sfe)
hvgs <- getTopHVGs(gvs, fdr.threshold = 0.05)
colGraph(sfe, "visium") <- findVisiumGraph(sfe)
sfe <- runMultivariate(sfe, "multispati", subset_row = hvgs)
}
