% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colFeatureData}
\alias{colFeatureData}
\alias{rowFeatureData}
\title{Get metadata of colData and rowData}
\usage{
colFeatureData(sfe)

rowFeatureData(sfe)
}
\arguments{
\item{sfe}{An SFE object.}
}
\value{
A \code{DataFrame}.
}
\description{
Results of spatial analyses on columns in \code{colData} and \code{rowData}
are stored in the metadata of \code{colData} and \code{rowData}, which can be
accessed by the \code{\link{metadata}} function. This function allows the
users to more directly access these results.
}
\examples{
library(SpatialFeatureExperiment)
library(SingleCellExperiment)
library(SFEData)
sfe <- McKellarMuscleData("small")
colGraph(sfe, "visium") <- findVisiumGraph(sfe)
# Moran's I for colData
sfe <- colDataMoransI(sfe, "nCounts")
colFeatureData(sfe)
}
