% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ditto_colors}
\alias{ditto_colors}
\title{Colorblind friendly palette from dittoSeq}
\format{
A character vector of hex colors of the palette. There are 40 colors.
}
\source{
The dittoSeq package.
}
\usage{
ditto_colors
}
\description{
Just to get the palette without having to install all those dependencies of
dittoSeq.
}
\keyword{datasets}
