% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGraphs.R
\name{findSpatialNeighbors,SpatialFeatureExperiment-method}
\alias{findSpatialNeighbors,SpatialFeatureExperiment-method}
\title{Find spatial neighborhood graph}
\usage{
\S4method{findSpatialNeighbors}{SpatialFeatureExperiment}(
  x,
  sample_id,
  geometry = "spatialCoords",
  MARGIN = 2,
  method = c("tri2nb", "knearneigh", "dnearneigh", "gabrielneigh", "relativeneigh",
    "soi.graph", "poly2nb"),
  glist = NULL,
  style = "W",
  zero.policy = NULL,
  ...
)
}
\arguments{
\item{x}{A \code{\link{SpatialFeatureExperiment}} object.}

\item{sample_id}{Which sample in the SFE object to use for the graph.}

\item{geometry}{Name of the geometry associated with the MARGIN of interest
for which to compute the graph.}

\item{MARGIN}{Just like in \code{\link{apply}}, where 1 stands for row, 2
stands for column. Here, in addition, 3 stands for annotation, to query the
\code{\link{annotGeometries}}, such as nuclei segmentation in a Visium data}

\item{method}{Name of function in the package \code{spdep} to use to find the
spatial neighborhood graph.}

\item{glist}{list of general weights corresponding to neighbours}

\item{style}{\code{style} can take values \dQuote{W}, \dQuote{B}, \dQuote{C}, \dQuote{U}, \dQuote{minmax} and \dQuote{S}}

\item{zero.policy}{default NULL, use global option value; if FALSE stop with error for any empty neighbour sets, if TRUE permit the weights list to be formed with zero-length weights vectors}

\item{...}{Extra arguments passed to the \code{spdep} function stated in the
\code{method} argument, such as \code{k}, \code{use_kd_tree}, \code{d1},
\code{d2}, \code{nnmult}, \code{sym}, and \code{quadsegs}. Note that any
arguments about using longitude and latitude, which are irrelevant, are
ignored. The \code{longlat} argument is hard coded to \code{FALSE}.}
}
\value{
A \code{listw} object representing the graph, with an attribute
  "method" recording the function used to build the graph, its arguments, and
  information about the geometry for which the graph was built. The attribute
  is used to reconstruct the graphs when the SFE object is subsetted since
  some nodes in the graph will no longer be present.
}
\description{
This function wraps all spatial neighborhood graphs implemented in the
package \code{spdep} for the \code{SpatialFeatureExperiment} (SFE) class, to
find spatial neighborhood graphs for the entities represented by columns or
rows of the gene count matrix in the SFE object or spatial entities in the
\code{annotGeometries} field of the SFE object. Results are stored as
\code{listw} objects in the \code{spatialGraphs} field of the SFE object, as
\code{listw} is used in many methods that facilitate the spatial neighborhood
graph in the \code{spdep}, \code{spatialreg}, and \code{adespatial}. The edge
weights of the graph in the \code{listw} object are by default style W (see
\code{\link{nb2listw}}) and the unweighted neighbor list is in the
\code{neighbours} field of the \code{listw} object.
}
