% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{mockSVGenes}
\alias{mockSVGenes}
\title{Generate a mock SpatialExperiment including spatially variable genes.}
\usage{
mockSVGenes(tot_genes, de_genes, lambda)
}
\arguments{
\item{tot_genes}{Total number of gene profiles to be generated.}

\item{de_genes}{Number of spatially variable genes.}

\item{lambda}{Expected number of spots over which genes are sampled.}
}
\value{
A \link{SpatialExperiment} object.
}
\description{
Implementation based on \code{\link[trendsceek:add_markdist_step]{trendsceek::add_markdist_step()}}.
}
\examples{
mockSVGenes(20, 2, 20)

}
