% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{moranPlot}
\alias{moranPlot}
\title{Use ggplot to plot the moran.plot results}
\usage{
moranPlot(
  sfe,
  feature,
  colGraphName,
  sample_id = NULL,
  contour_color = "cyan",
  color_by = NULL,
  colGeometryName = NULL,
  annotGeometryName = NULL,
  plot_singletons = TRUE,
  filled = FALSE,
  divergent = FALSE,
  diverge_center = NULL,
  name = "MoranPlot",
  ...
)
}
\arguments{
\item{sfe}{A \code{SpatialFeatureExperiment} object.}

\item{feature}{Name of one variable to show on the plot. It will be converted
to sentence case on the x axis and lower case in the y axis appended after
"Spatially lagged". One feature at a time since the colors in
\code{color_by} may be specific to this feature (e.g. from
\code{\link{clusterMoranPlot}}).}

\item{colGraphName}{Name of the \code{colGraph}, the spatial neighborhood
graph used to compute the Moran plot. This is to determine which points are
singletons to plot differently on this plot.}

\item{sample_id}{One sample_id for the sample whose graph to plot.}

\item{contour_color}{Color of the point density contours, which can be
changed so the contours stand out from the points.}

\item{color_by}{Variable to color the points by. It can be the name of a
column in colData, a gene, or the name of a column in the colGeometry
specified in colGeometryName. Or it can be a vector of the same length as
the number of cells/spots in the sample_id of interest.}

\item{colGeometryName}{Name of a \code{colGeometry} \code{sf} data frame
whose numeric columns of interest are to be used to compute the metric. Use
\code{\link{colGeometryNames}} to look up names of the \code{sf} data
frames associated with cells/spots.}

\item{annotGeometryName}{Name of a \code{annotGeometry} of the SFE object, to
annotate the gene expression plot.}

\item{plot_singletons}{Logical, whether to plot items that don't have spatial
neighbors.}

\item{filled}{Logical, whether to plot filled contours for the
non-influential points and only plot influential points as points.}

\item{divergent}{Logical, whether a divergent palette should be used.}

\item{diverge_center}{If \code{divergent = TRUE}, the center from which the
palette should diverge. If \code{NULL}, then not centering.}

\item{name}{String specifying the name to be used to store the results in
\code{rowData(x)}. If not already present in the name, then the
\code{sample_id} will be appended to the name specified here separated by
an underscore.}

\item{...}{Other arguments to pass to \code{\link{geom_density2d}}.}
}
\value{
A ggplot object.
}
\description{
This function uses \code{ggplot2} to plot the Moran plot. The plot would be
more aesthetically pleasing than the base R version implemented in
\code{spdep}. In addition, contours are plotted to show point density on the
plot, and the points can be colored by a variable, such as clusters. The
contours may also be filled and only influential points plotted. When filled,
the viridis E option is used.
}
