% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCellBin2D}
\alias{plotCellBin2D}
\title{Plot cell density as 2D histogram}
\usage{
plotCellBin2D(sfe, bins = 200, binwidth = NULL, hex = FALSE)
}
\arguments{
\item{sfe}{A \code{SpatialFeatureExperiment} object.}

\item{bins}{Numeric vector giving number of bins in both vertical and
horizontal directions. Set to 100 by default.}

\item{binwidth}{Numeric vector giving bin width in both vertical and
horizontal directions. Overrides \code{bins} if both set.}

\item{hex}{Logical, whether to use hexagon rather than rectangular bins.
Requires the \code{hexbin} package.}
}
\value{
A ggplot object.
}
\description{
This function plots cell density in histological space as 2D histograms, 
especially helpful for larger smFISH-based datasets.
}
\examples{
library(SFEData)
sfe <- HeNSCLCData()
plotCellBin2D(sfe)
}
