% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-non-spatial.R
\name{plotColDataHistogram}
\alias{plotColDataHistogram}
\alias{plotRowDataHistogram}
\title{Plot histograms for colData and rowData columns}
\usage{
plotColDataHistogram(
  sfe,
  feature,
  fill_by = NULL,
  subset = NULL,
  bins = 100,
  binwidth = NULL,
  scales = "free",
  ncol = 1,
  position = "identity"
)

plotRowDataHistogram(
  sfe,
  feature,
  fill_by = NULL,
  subset = NULL,
  bins = 100,
  binwidth = NULL,
  scales = "free",
  ncol = 1,
  position = "identity"
)
}
\arguments{
\item{sfe}{A \code{SpatialFeatureExperiment} object.}

\item{feature}{Names of columns in \code{colData} or \code{rowData} to plot.
When multiple features are specified, they will be plotted in separate
facets.}

\item{fill_by}{Name of a categorical column in \code{colData} or
\code{rowData} to fill the histogram.}

\item{subset}{Name of a logical column to only plot a subset of the data.}

\item{bins}{Numeric vector giving number of bins in both vertical and
horizontal directions. Set to 100 by default.}

\item{binwidth}{Numeric vector giving bin width in both vertical and
horizontal directions. Overrides \code{bins} if both set.}

\item{scales}{Should scales be fixed (\code{"fixed"}, the default),
free (\code{"free"}), or free in one dimension (\code{"free_x"},
\code{"free_y"})?}

\item{ncol}{Number of columns in the facetting.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}
}
\value{
A ggplot object
}
\description{
Plot histograms for colData and rowData columns
}
\examples{
library(SFEData)
sfe <- McKellarMuscleData()
plotColDataHistogram(sfe, c("nCounts", "nGenes"), fill_by = "in_tissue",
                     bins = 50)
plotColDataHistogram(sfe, "nCounts", subset = "in_tissue")
sfe2 <- sfe[, sfe$in_tissue]
plotColDataHistogram(sfe2, c("nCounts", "nGenes"), bins = 50)
}
