% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotColGraph}
\alias{plotColGraph}
\alias{plotAnnotGraph}
\title{Plot spatial graphs}
\usage{
plotColGraph(
  sfe,
  colGraphName = 1L,
  colGeometryName = NULL,
  sample_id = NULL,
  segment_size = 0.5,
  geometry_size = 0.5,
  ncol = NULL
)

plotAnnotGraph(
  sfe,
  annotGraphName = 1L,
  annotGeometryName = 1L,
  sample_id = NULL,
  segment_size = 0.5,
  geometry_size = 0.5,
  ncol = NULL
)
}
\arguments{
\item{sfe}{A \code{SpatialFeatureExperiment} object.}

\item{colGraphName}{Name of graph associated with columns of the gene count
matrix to be plotted.}

\item{colGeometryName}{Name of a \code{colGeometry} \code{sf} data frame
whose numeric columns of interest are to be used to compute the metric. Use
\code{\link{colGeometryNames}} to look up names of the \code{sf} data
frames associated with cells/spots.}

\item{sample_id}{Sample(s) in the SFE object whose cells/spots to use. Can be
"all" to compute metric for all samples; the metric is computed separately
for each sample.}

\item{segment_size}{Thickness of the segments that represent graph edges.}

\item{geometry_size}{Point size (for POINT geometries) or line thickness (for
LINESTRING and POLYGON) to plot the geometry in the background.}

\item{ncol}{Number of columns if plotting multiple features. Defaults to
\code{NULL}, which means using the same logic as \code{facet_wrap}, which
is used by \code{patchwork}'s \code{\link{wrap_plots}} by default.}

\item{annotGraphName}{Name of the annotation graph to plot.}

\item{annotGeometryName}{Name of the \code{annotGeometry}, which is
associated with the graph specified with \code{annotGraphName}, for spatial
coordinates of the graph nodes and for context.}
}
\value{
A ggplot2 object.
}
\description{
A ggplot version of \code{spdep::plot.nb}, reducing boilerplate for SFE
objects.
}
