% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCorrelogram}
\alias{plotCorrelogram}
\title{Plot correlogram}
\usage{
plotCorrelogram(
  sfe,
  features,
  sample_id = NULL,
  method = "I",
  color_by = NULL,
  colGeometryName = NULL,
  annotGeometryName = NULL,
  plot_signif = TRUE,
  p_adj_method = "BH",
  divergent = FALSE,
  diverge_center = NULL,
  name = paste("Correlogram", method, sep = "_")
)
}
\arguments{
\item{sfe}{A \code{SpatialFeatureExperiment} object.}

\item{features}{Features to plot, must be in rownames of the gene count
matrix, colnames of colData or a colGeometry.}

\item{sample_id}{Sample in the SFE object whose cells/spots to use.}

\item{method}{"corr" for correlation, "I" for Moran's I, "C" for Geary's C}

\item{color_by}{Name of a column in \code{rowData(sfe)} or in the
\code{featureData} attribute of \code{colData}, \code{colGeometry}, or
\code{annotGeometry} by which to color the correlogram of each feature.}

\item{colGeometryName}{Name of a \code{colGeometry} \code{sf} data frame
whose numeric columns of interest are to be used to compute the metric. Use
\code{\link{colGeometryNames}} to look up names of the \code{sf} data
frames associated with cells/spots.}

\item{annotGeometryName}{Name of a \code{annotGeometry} of the SFE object, to
annotate the gene expression plot.}

\item{plot_signif}{Logical, whether to plot significance symbols: p < 0.001:
***, p < 0.01: **, p < 0.05 *, p < 0.1: ., otherwise no symbol. The
p-values are two sided, based on the assumption that the estimated Moran's
I is normally distributed with mean from a randomized version of the data.
The mean and variance come from \code{\link{moran.test}} for Moran's I and
\code{\link{geary.test}} for Geary's C. Take the results with a grain of
salt if the data is not normally distributed.}

\item{p_adj_method}{Multiple testing correction method as in
\code{\link{p.adjust}}, to correct for multiple testing (number of lags
times number of features) in the Moran's I estimates if \code{plot_signif =
TRUE}.}

\item{divergent}{Logical, whether a divergent palette should be used.}

\item{diverge_center}{If \code{divergent = TRUE}, the center from which the
palette should diverge. If \code{NULL}, then not centering.}

\item{name}{String specifying the name to be used to store the results in
\code{rowData(x)}. If not already present in the name, then the
\code{sample_id} will be appended to the name specified here separated by
an underscore.}
}
\value{
A ggplot object.
}
\description{
Use \code{ggplot2} to plot correlograms computed by
\code{\link{runCorrelogram}}, pulling results from \code{rowData}.
Correlograms of multiple genes with error bars can be plotted, and they can
be colored by any numeric or categorical column in \code{rowData} or a vector
with the same length as \code{nrow} of the SFE object. The coloring is useful
when the correlograms are clustered to show types of length scales or
patterns of decay of spatial autocorrelation. For \code{method = "I"}, the
error bars are twice the standard deviation of the estimated Moran's I value.
}
