% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMoranMC}
\alias{plotMoranMC}
\title{Plot Moran/Geary monte carlo results}
\usage{
plotMoranMC(
  sfe,
  features,
  sample_id = NULL,
  facet_by = c("sample_id", "features"),
  ncol = NULL,
  colGeometryName = NULL,
  annotGeometryName = NULL,
  name = "moran.mc",
  ptype = c("density", "histogram", "freqpoly"),
  show_symbol = TRUE,
  ...
)
}
\arguments{
\item{sfe}{A \code{SpatialFeatureExperiment} object.}

\item{features}{Features to plot, must be in rownames of the gene count
matrix, colnames of colData or a colGeometry.}

\item{sample_id}{Sample(s) in the SFE object whose cells/spots to use. Can be
"all" to compute metric for all samples; the metric is computed separately
for each sample.}

\item{facet_by}{Whether to facet by sample_id (default) or features. If
facetting by sample_id, then different features will be plotted in the same
facet for comparison. If facetting by features, then different samples will
be compared for each feature. Ignored if only one sample is specified.}

\item{ncol}{Number of columns if facetting.}

\item{colGeometryName}{Name of a \code{colGeometry} \code{sf} data frame
whose numeric columns of interest are to be used to compute the metric. Use
\code{\link{colGeometryNames}} to look up names of the \code{sf} data
frames associated with cells/spots.}

\item{annotGeometryName}{Name of a \code{annotGeometry} of the SFE object, to
annotate the gene expression plot.}

\item{ptype}{Plot type, one of "density", "histogram", or "freqpoly".}

\item{show_symbol}{Logical, whether to show human readable gene symbol on the
plot instead of Ensembl IDs when the row names are Ensembl IDs. There must
be a column in \code{rowData(sfe)} called "symbol" for this to work.}

\item{...}{Other arguments passed to \code{\link{geom_density}},
\code{\link{geom_histogram}}, or \code{\link{geom_freqpoly}}, depending on
\code{ptype}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot the simulations as a density plot or histogram compared to the observed
Moran's I or Geary's C, with ggplot2 so it looks nicer. Unlike the plotting
function in \code{spdep}, this function can also plot the same feature in
different samples as facets or plot different features or samples together
for comparison.
}
\examples{
library(SpatialFeatureExperiment)
library(SFEData)
sfe <- McKellarMuscleData("small")
colGraph(sfe, "visium") <- findVisiumGraph(sfe)
sfe <- colDataMoranMC(sfe, "nCounts", nsim = 100)
plotMoranMC(sfe, "nCounts")
}
