% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSpatialFeature}
\alias{plotSpatialFeature}
\title{Plot gene expression in space}
\usage{
plotSpatialFeature(
  sfe,
  colGeometryName,
  features,
  sample_id = NULL,
  ncol = NULL,
  annotGeometryName = NULL,
  annot_aes = list(),
  annot_fixed = list(),
  exprs_values = "logcounts",
  aes_use = c("fill", "color", "shape", "linetype"),
  divergent = FALSE,
  diverge_center = NULL,
  annot_divergent = FALSE,
  annot_diverge_center = NULL,
  only_plot_expressed = FALSE,
  size = 0,
  shape = 16,
  linetype = 1,
  alpha = 1,
  color = "black",
  fill = "gray80",
  ...
)
}
\arguments{
\item{sfe}{A \code{SpatialFeatureExperiment} object.}

\item{colGeometryName}{Name of a \code{colGeometry} \code{sf} data frame
whose numeric columns of interest are to be used to compute the metric. Use
\code{\link{colGeometryNames}} to look up names of the \code{sf} data
frames associated with cells/spots.}

\item{features}{Features to plot, must be in rownames of the gene count
matrix, colnames of colData or a colGeometry.}

\item{sample_id}{Sample(s) in the SFE object whose cells/spots to use. Can be
"all" to compute metric for all samples; the metric is computed separately
for each sample.}

\item{ncol}{Number of columns if plotting multiple features. Defaults to
\code{NULL}, which means using the same logic as \code{facet_wrap}, which
is used by \code{patchwork}'s \code{\link{wrap_plots}} by default.}

\item{annotGeometryName}{Name of a \code{annotGeometry} of the SFE object, to
annotate the gene expression plot.}

\item{annot_aes}{A named list of plotting parameters for the annotation sf
data frame. The names are which geom (as in ggplot2, such as color and
fill), and the values are column names in the annotation sf data frame.
Tidyeval is NOT supported.}

\item{annot_fixed}{Similar to \code{annot_aes}, but for fixed aesthetic
settings, such as \code{color = "gray"}. The defaults are the same as the
relevant defaults for this function.}

\item{exprs_values}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{aes_use}{Aesthetic to use for discrete variables. For continuous
variables, it's always "fill" for polygons and point shapes 21-25. For
discrete variables, it can be fill, color, shape, or linetype, whenever
applicable. The specified value will be changed to the applicable
equivalent. For example, if the geometry is point but "linetype" is
specified, then "shaped" will be used instead.}

\item{divergent}{Logical, whether a divergent palette should be used.}

\item{diverge_center}{If \code{divergent = TRUE}, the center from which the
palette should diverge. If \code{NULL}, then not centering.}

\item{annot_divergent}{Just as \code{divergent}, but for the annotGeometry in
case it's different.}

\item{annot_diverge_center}{Just as \code{diverge_center}, but for the
annotGeometry in case it's different.}

\item{only_plot_expressed}{Logical, whether to only plot values > 0. This
argument is only used when all values are non-negative.}

\item{size}{Fixed size of points or width of lines, including outlines of
polygons. For polygons, this defaults to 0, meaning no outlines. For points
and lines, this defaults to 0.5. Ignored if \code{size_by} is specified.}

\item{shape}{Fixed shape of points, ignored if \code{shape_by} is specified
and applicable.}

\item{linetype}{Fixed line type, ignored if \code{linetype_by} is specified
and applicable.}

\item{alpha}{Transparency.}

\item{color}{Fixed color for \code{colGeometry} if \code{color_by} is not
specified or not applicable, or for \code{annotGeometry} if
\code{annot_color_by} is not specified or not applicable.}

\item{fill}{Similar to \code{color}, but for fill.}

\item{...}{Other arguments passed to \code{\link{wrap_plots}}.}
}
\description{
Unlike \code{Seurat} and \code{ggspavis}, plotting functions in this package
uses \code{geom_sf} whenever applicable.
}
\details{
In the documentation of this function, a "feature" can be a gene (or whatever
entity that corresponds to rows of the gene count matrix), a column in
\code{colData}, or a column in the \code{colGeometry} \code{sf} data frame
specified in the \code{colGeometryName} argument.

For continuous variables, the Blues palette from colorbrewer is used if
\code{divergent = FALSE}, and the roma palette from the \code{scico} package
if \code{divergent = TRUE}. For discrete variables, the \code{dittoSeq}
palette is used. The defaults are colorblind friendly. For annotation, the
PuRd colorbrewer palette is used for continuous variables and the other end
of the \code{dittoSeq} palette is used for discrete variables.
}
