% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFEMethod-class.R
\docType{class}
\name{SFEMethod}
\alias{SFEMethod}
\alias{SFEMethod-class}
\alias{info,SFEMethod-method}
\alias{is_local,SFEMethod-method}
\alias{fun,SFEMethod-method}
\alias{to_df_fun,SFEMethod-method}
\title{SFEMethod class}
\usage{
SFEMethod(info, fun, to_df_fun)

\S4method{info}{SFEMethod}(x, type)

\S4method{is_local}{SFEMethod}(x)

\S4method{fun}{SFEMethod}(x)

\S4method{to_df_fun}{SFEMethod}(x)
}
\arguments{
\item{info}{See slot documentation}

\item{fun}{See slot documentation}

\item{to_df_fun}{See slot documentation}

\item{x}{A \code{SFEMethod} object}

\item{type}{One of the names of the \code{info} slot, see slot documentation.}
}
\value{
The constructor returns a \code{SFEMethod} object. The getters return
the content of the corresponding slots.
}
\description{
This S4 class is used to wrap spatial analysis methods, taking inspiration
from the \code{caret} and \code{tidymodels} packages.
}
\section{Slots}{

\describe{
\item{\code{info}}{A named character vector specifying information about the method:
\describe{
\item{name}{Name of the method, used by user-facing functions to specify the
method to use, such as "moran" for Moran's I.}
\item{variate}{How many variables this method works with, must be one of
"uni" for univariate, "bi" for bivariate, or "multi" for multivariate.}
\item{scope}{Either "global", returning one result for tne entire dataset,
or "local", returning one result for each spatial location.}
\item{package}{Name of the package whose implementation of the method is used
here, used to check if the package is installed.}
\item{title}{Descriptive title to show when plotting the results.}
\item{default_attr}{For local methods that return multiple fields, such as
local Moran values and their p-values, the default field to use when
plotting.}
}}

\item{\code{fun}}{The function implementing the method. For univariate methods, the
first two arguments must be \code{x} for a vector, and \code{listw} for a
\code{listw} object specifying the spatial neighborhood graph, and
optionally \code{...}. If the original function implementing the method in
the package has different argument names or orders, write a thin wrapper to
rearrange and/or rename the arguments.}

\item{\code{to_df_fun}}{Function to convert output from \code{fun} into a format to
store in the SFE object. For univariate global methods, different fields of
the result should be columns of a data frame with one row so results for
multiple features will be a data frame. The arguments should be \code{out}
for a list of raw output, each element of which is output for one feature
and \code{name} to rename the primary field if a more informative name is
needed, and \code{...} for other arguments specific to methods. For
univariate local methods, the output should be a data frame or matrix whose
rows match the columns of the gene count matrix. The arguments should be
\code{out}, \code{nb} for a neighborhood list used for multiple testing
correction, and \code{p.adjust.method} for a method to correct for multiple
testing as in \code{\link{p.adjust}}, and \code{...}.}
}}

