% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpatialReducedDim.R
\name{spatialReducedDim}
\alias{spatialReducedDim}
\title{Plot dimension reduction components in space}
\usage{
spatialReducedDim(
  sfe,
  dimred,
  ncomponents,
  colGeometryName = 1L,
  sample_id = "all",
  ncol = NULL,
  ncol_sample = NULL,
  annotGeometryName = NULL,
  annot_aes = list(),
  annot_fixed = list(),
  exprs_values = "logcounts",
  bbox = NULL,
  aes_use = c("fill", "color", "shape", "linetype"),
  divergent = FALSE,
  diverge_center = NULL,
  annot_divergent = FALSE,
  annot_diverge_center = NULL,
  size = 0,
  shape = 16,
  linetype = 1,
  alpha = 1,
  color = NA,
  fill = "gray80",
  scattermore = FALSE,
  pointsize = 0,
  ...
)
}
\arguments{
\item{sfe}{A \code{SpatialFeatureExperiment} object.}

\item{dimred}{A string or integer scalar indicating the reduced dimension
result in \code{reducedDims(sfe)} to plot.}

\item{ncomponents}{A numeric scalar indicating the number of dimensions to
plot, starting from the first dimension. Alternatively, a numeric vector
specifying the dimensions to be plotted.}

\item{colGeometryName}{Name of a \code{colGeometry} \code{sf} data frame
whose numeric columns of interest are to be used to compute the metric. Use
\code{\link{colGeometryNames}} to look up names of the \code{sf} data
frames associated with cells/spots.}

\item{sample_id}{Sample(s) in the SFE object whose cells/spots to use. Can be
"all" to compute metric for all samples; the metric is computed separately
for each sample.}

\item{ncol}{Number of columns if plotting multiple features. Defaults to
\code{NULL}, which means using the same logic as \code{facet_wrap}, which
is used by \code{patchwork}'s \code{\link{wrap_plots}} by default.}

\item{ncol_sample}{If plotting multiple samples as facets, how many columns
of such facets. This is distinct from \code{ncols}, which is for multiple
features. When plotting multiple features for multiple samples, then the
result is a multi-panel plot each panel of which is a plot for each feature
facetted by samples.}

\item{annotGeometryName}{Name of a \code{annotGeometry} of the SFE object, to
annotate the gene expression plot.}

\item{annot_aes}{A named list of plotting parameters for the annotation sf
data frame. The names are which geom (as in ggplot2, such as color and
fill), and the values are column names in the annotation sf data frame.
Tidyeval is NOT supported.}

\item{annot_fixed}{Similar to \code{annot_aes}, but for fixed aesthetic
settings, such as \code{color = "gray"}. The defaults are the same as the
relevant defaults for this function.}

\item{exprs_values}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{bbox}{A bounding box to specify a smaller region to plot, useful when
the dataset is large. All geometries that intersect the bounding box will
be plotted. Can be a named numeric vector with names "xmin", "xmax",
"ymin", and "ymax", in any order. If plotting multiple samples, it should
be a matrix with sample IDs as column names and "xmin", "ymin", "xmax", and
"ymax" as row names. If multiple samples are plotted but \code{bbox} is a
vector rather than a matrix, then the same bounding box will be used for
all samples. If \code{NULL}, then the entire tissue is plotted.}

\item{aes_use}{Aesthetic to use for discrete variables. For continuous
variables, it's always "fill" for polygons and point shapes 21-25. For
discrete variables, it can be fill, color, shape, or linetype, whenever
applicable. The specified value will be changed to the applicable
equivalent. For example, if the geometry is point but "linetype" is
specified, then "shaped" will be used instead.}

\item{divergent}{Logical, whether a divergent palette should be used.}

\item{diverge_center}{If \code{divergent = TRUE}, the center from which the
palette should diverge. If \code{NULL}, then not centering.}

\item{annot_divergent}{Just as \code{divergent}, but for the annotGeometry in
case it's different.}

\item{annot_diverge_center}{Just as \code{diverge_center}, but for the
annotGeometry in case it's different.}

\item{size}{Fixed size of points or width of lines, including outlines of
polygons. For polygons, this defaults to 0, meaning no outlines. For points
and lines, this defaults to 0.5. Ignored if \code{size_by} is specified.}

\item{shape}{Fixed shape of points, ignored if \code{shape_by} is specified
and applicable.}

\item{linetype}{Fixed line type, ignored if \code{linetype_by} is specified
and applicable.}

\item{alpha}{Transparency.}

\item{color}{Fixed color for \code{colGeometry} if \code{color_by} is not
specified or not applicable, or for \code{annotGeometry} if
\code{annot_color_by} is not specified or not applicable.}

\item{fill}{Similar to \code{color}, but for fill.}

\item{scattermore}{Logical, whether to use the \code{scattermore} package to
greatly speed up plotting numerous points. Only used for POINT
\code{colGeometries}. If the geometry is not POINT, then the centroids are
used. Recommended for plotting hundreds of thousands or more cells where
the cell polygons can't be seen when plotted due to the large number of
cells and small plot size such as when plotting multiple panels for
multiple features.}

\item{pointsize}{Radius of rasterized point in \code{scattermore}. Default to
0 for single pixels (fastest).}

\item{...}{Other arguments passed to \code{\link{wrap_plots}}.}
}
\value{
Same as in \code{\link{plotSpatialFeature}}. A \code{ggplot2} object
  if plotting one component. A \code{patchwork} object if plotting multiple
  components.
}
\description{
Such as plotting the value of projection of gene expression of each cell to a
principal component in space. At present, this function does not work for the
3D array of geographically weighted PCA (GWPCA), but a future version will
deal with GWPCA results.
}
\examples{
library(SFEData)
library(scater)
sfe <- McKellarMuscleData("small")
sfe <- logNormCounts(sfe)
sfe <- runPCA(sfe, ncomponents = 2)
spatialReducedDim(sfe, "PCA", 2, "spotPoly",
    annotGeometryName = "tissueBoundary",
    divergent = TRUE, diverge_center = 0
)
# Basically PC1 separates spots not on tissue from those on tissue.
}
\seealso{
scater::plotReducedDim
}
