% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.R
\docType{methods}
\name{svGenes}
\alias{svGenes}
\alias{svGenes,SpatialExperiment-method}
\alias{svGenes,SpatialExperiment,ANY-method}
\title{Identifies spatially variable genes.}
\usage{
svGenes(x, method, ...)

\S4method{svGenes}{SpatialExperiment}(x, method = c("spatialde", "spark"), assay = "counts")
}
\arguments{
\item{x}{SpatialExpression object.}

\item{method}{Name of the method for testing spatial variabily.}

\item{...}{Additional arguments.}

\item{assay}{Extract gene expression from this assay.}
}
\value{
Extended object with DE information.

Extended SpatialExpression object.
}
\description{
You can select one among the following statistical tests:
\itemize{
\item spatialde, from \href{https://www.nature.com/articles/nmeth.4636}{Svensson et al, 2018}.
\item spark, from \href{https://www.nature.com/articles/s41592-019-0701-7}{Sun et al, 2020}.
}
}
\examples{
svGenes(mockSVGenes(100, 4, 40), method = "spatialde")

}
