% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VoyageR.R
\name{VoyageR}
\alias{VoyageR}
\title{Launch Voyager as a Shiny Gadget}
\usage{
VoyageR(
  data = NULL,
  format = "vegawidget",
  browser = FALSE,
  envir = .GlobalEnv
)
}
\arguments{
\item{data}{Dataset to load}

\item{format}{In what format should plots be returned: "vegawidget" to return
an interactive or rendered version, "json" to return the spec as a character
in json format}

\item{browser}{Open in an external browser}

\item{envir}{Environment to use when searching for datasets
(defaults to the global environment)}
}
\value{
If a visualisation has been specified its JSON will be returned as a string
}
\description{
Launch Voyager as a Shiny Gadget
}
\examples{
data("mtcars")

# Don't run the app when testing examples
if (interactive()) {
  # Launch app and return a vegawidget
  VoyageR(mtcars)
}

if (interactive()) {
  # Return the vega plot specification as json
  spec_json <- VoyageR(mtcars, format = "json")
  vegawidget::as_vegaspec(spec_json)
}
}
